Service editing: some help for Mendeley and AuthorClaim
I added input prepend and append (using bootstrap css classes
- may need to be changed/added to site's css) to show as a URL
pattern (which constructs the links further up in this file),
as well as a bit of descriptive text on where to find the
desired information.

I didn't actually add any CSS or test this beyond fooling
around in the WebKit/Chrome Inspector and adding this, this
is just I would expect to work. This is almost more of an update
to issue #19 by example rather than code I'd expect you to directly merg.

Refactoring: could use DRY for the URL prefix/suffix, and really I'd think that
those repetitive elsifs for the services could probably be factored out, but I'm
not too experienced with ruby and have less experience with Rails.